+++ utils/trainer_util.py
@@ -5,6 +5,6 @@ class TrainerUtil:
     @staticmethod
     def get_sigmoid_correct_count(predict, tlabel):
         predict = torch.round(predict)
-        diff = tlabel.float().sub(predict)
+        diff = tlabel.sub(predict)
         abs = torch.abs(diff)
         return predict.size()[0] - torch.sum(abs).item()
